+++ docs/news.rst
@@ -13,7 +13,7 @@ Future releases
 Psycopg 3.0.14 (unreleased)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-- Fail dumping arrays of mixed numbers with DataError (:ticket:`#301`).
+- Raise DataError dumping arrays of mixed types (:ticket:`#301`).
 
 
 Current release